backend/requirements.txt
django
django-environ
djangorestframework
psycopg2-binary
django-cors-headers
environ
drf_spectacular